+++ install/migrations/update_10.0.x_to_10.1.0/notifications.php
@@ -288,3 +288,32 @@ if (countElementsInTable('glpi_notifications', ['itemtype' => 'Ticket', 'event' 
     ]);
 }
 /** /Add new notification for AutoBump */
+
+/** Add handling of source item of attached documents in notification */
+$migration->addField("glpi_notifications", "attach_documents", "tinyint NOT NULL DEFAULT '-2'");
+
+$attach_documents_value = $DB->request([
+    'SELECT' => ['value'],
+    'FROM'   => 'glpi_configs',
+    'WHERE'  => [
+        'context' => 'core',
+        'name'    => 'attach_ticket_documents_to_mail'
+    ],
+])->current()['value'] ?? '0';
+$migration->addField(
+    "glpi_queuednotifications",
+    "attach_documents",
+    "tinyint NOT NULL DEFAULT '0'",
+    [
+        'update' => $attach_documents_value,
+    ]
+);
+
+$migration->addField("glpi_queuednotifications", "itemtype_trigger", "varchar(255) DEFAULT NULL");
+$migration->addField("glpi_queuednotifications", "items_id_trigger", "fkey");
+$migration->addKey(
+    "glpi_queuednotifications",
+    ["itemtype_trigger", "items_id_trigger"],
+    "item_trigger"
+);
+/** /Add handling of source item of attached documents in notification */