+++ install/mysql/glpi-empty.sql
@@ -4954,6 +4954,7 @@ CREATE TABLE `glpi_notifications` (
   `date_mod` timestamp NULL DEFAULT NULL,
   `date_creation` timestamp NULL DEFAULT NULL,
   `allow_response` tinyint NOT NULL DEFAULT '1',
+  `attach_documents` tinyint NOT NULL DEFAULT '-2',
   PRIMARY KEY (`id`),
   KEY `name` (`name`),
   KEY `itemtype` (`itemtype`),
@@ -6173,6 +6174,9 @@ CREATE TABLE `glpi_queuednotifications` (
   `documents` text,
   `mode` varchar(20) NOT NULL COMMENT 'See Notification_NotificationTemplate::MODE_* constants',
   `event` varchar(255) DEFAULT NULL,
+  `attach_documents` tinyint NOT NULL DEFAULT '0',
+  `itemtype_trigger` varchar(255) DEFAULT NULL,
+  `items_id_trigger` int unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
   KEY `item` (`itemtype`,`items_id`,`notificationtemplates_id`),
   KEY `is_deleted` (`is_deleted`),
@@ -6182,7 +6186,8 @@ CREATE TABLE `glpi_queuednotifications` (
   KEY `send_time` (`send_time`),
   KEY `sent_time` (`sent_time`),
   KEY `mode` (`mode`),
-  KEY `notificationtemplates_id` (`notificationtemplates_id`)
+  KEY `notificationtemplates_id` (`notificationtemplates_id`),
+  KEY `item_trigger` (`itemtype_trigger`,`items_id_trigger`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci ROW_FORMAT=DYNAMIC;